chore: remove verbosity hack for `polycli monitor`
